split constraint definitions into separate files
